requirements.txt
asgiref==3.7.2
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-nested-routers==0.93.4
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
flake8==6.1.0
